# site/src/data/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/tandem/readiness": {
      "get": {
        "description": "Same evaluation as `GET /ready`, returned under the protocol prefix. Every failing gate is listed in `reasons`.",
        "operationId": "TandemController_readinessSnapshot",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Every readiness gate passed."
          },
          "503": {
            "description": "At least one gate failed. The body is the full snapshot including `reasons`."
          }
        },
        "summary": "Readiness snapshot with the reasons behind it",
        "tags": [
          "tandem"
        ]
      }
    },
    "/tandem/objects/{objectKey}": {
      "get": {
        "description": "Looks up an object by its 32-byte binary key and returns its current state together with every chapter in sequence order.",
        "operationId": "TandemController_object",
        "parameters": [
          {
            "name": "objectKey",
            "required": true,
            "in": "path",
            "description": "The 32-byte object key as 64 lowercase hexadecimal characters.",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "The object and its ordered chapters."
          },
          "400": {
            "description": "`object key must be 32-byte hex`."
          },
          "404": {
            "description": "`object not found`."
          }
        },
        "summary": "One object and its chapters",
        "tags": [
          "tandem"
        ]
      }
    },
    "/tandem/carriers/{txid}/{vout}": {
      "get": {
        "description": "Returns the carrier record for one outpoint, including the height that created it and the transaction that spent it if it is no longer current.",
        "operationId": "TandemController_carrier",
        "parameters": [
          {
            "name": "vout",
            "required": true,
            "in": "path",
            "description": "Output index. A carrier is always output index 1.",
            "schema": {}
          },
          {
            "name": "txid",
            "required": true,
            "in": "path",
            "description": "Transaction id in the usual lowercase display order, 64 hexadecimal characters.",
            "schema": {
              "example": "9d2f0c1a4b6e8d3f5a7c9e1b3d5f7a9c1e3b5d7f9a1c3e5b7d9f1a3c5e7b9d1f"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The carrier record."
          },
          "400": {
            "description": "`txid must be 32-byte hex` or `vout is invalid`."
          },
          "404": {
            "description": "`carrier not found`."
          }
        },
        "summary": "Carrier state at an exact outpoint",
        "tags": [
          "tandem"
        ]
      }
    },
    "/tandem/events/{txid}": {
      "get": {
        "description": "Returns every event this transaction produced, ordered by event index. A known transaction that produced no event returns an empty list rather than a 404.",
        "operationId": "TandemController_events",
        "parameters": [
          {
            "name": "txid",
            "required": true,
            "in": "path",
            "description": "Transaction id in the usual lowercase display order, 64 hexadecimal characters.",
            "schema": {
              "example": "9d2f0c1a4b6e8d3f5a7c9e1b3d5f7a9c1e3b5d7f9a1c3e5b7d9f1a3c5e7b9d1f"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The transaction id and its ordered events."
          },
          "400": {
            "description": "`txid must be 32-byte hex`."
          },
          "404": {
            "description": "`transaction not found`."
          }
        },
        "summary": "Canonical events for one transaction",
        "tags": [
          "tandem"
        ]
      }
    },
    "/tandem/invalid-events": {
      "get": {
        "description": "Validity class 0 events, newest first. These record a rejected observation and its reason code without mutating any object.",
        "operationId": "TandemController_invalidEvents",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Row cap between 1 and 200. Anything outside that range is rejected.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invalid observations, newest first."
          }
        },
        "summary": "Observations that produced no state",
        "tags": [
          "tandem"
        ]
      }
    },
    "/tandem/reorgs": {
      "get": {
        "description": "One row per completed rollback, newest first, recording the old tip, the common ancestor, the new tip, and how many blocks were rolled back.",
        "operationId": "TandemController_reorgs",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Row cap between 1 and 200. Anything outside that range is rejected.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Completed rollbacks, newest first."
          }
        },
        "summary": "Reorganization journal",
        "tags": [
          "tandem"
        ]
      }
    },
...
    "/tandem/agreement/{height}": {
      "get": {
        "description": "Returns the Ed25519 envelope over the RFC 8785 canonical JSON of the tuple for one canonical height. This is the endpoint an independent pipeline calls to obtain a second opinion, and the same shape this pipeline expects from pipeline B.",
        "operationId": "TandemController_agreement",
        "parameters": [
          {
            "name": "height",
            "required": true,
            "in": "path",
            "description": "Canonical block height.",
            "schema": {
              "example": 1008
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The signed agreement envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgreementEnvelope"
                }
              }
            }
          }
        },
        "summary": "This pipeline's signed agreement tuple at a height",
        "tags": [
          "tandem"
        ]
      }
    },
...
  },
  "info": {
    "title": "Tandem indexer pipeline A",
    "description": "Tandem is a Bitcoin object held jointly by exactly two keys. This service indexes one deployment of it and exposes two separate HTTP surfaces.\n\n`/tandem` is the direct surface. It answers from this pipeline's own view and makes no claim that any other implementation agrees.\n\n`/tandem/verified` is the fail-closed surface. Before returning any data it requires a signed agreement tuple from this pipeline and an independently signed tuple from pipeline B at the same canonical height, verifies both signatures against configured trust maps, and compares nine semantic fields. It verifies again after the read. If anything is missing, stale, untrusted, unverifiable, or different, it returns HTTP 503 rather than an answer it cannot stand behind.\n\nPipeline A is not an authority for wallet spending. A missing, stale, or disagreeing tuple must block mint and protected-output spending flows.\n\nMainnet responses on the verified surface are refused unless `TANDEM_VERIFIED_MAINNET_ENABLED=true` is set deliberately.",
    "version": "0.1.0",
    "contact": {},
    "license": {
      "name": "MIT",
      "url": "https://github.com/bitcoinuniverse/index-tandem/blob/main/LICENSE"
    }
  },
...
  "servers": [
    {
      "url": "http://127.0.0.1:3021",
      "description": "Local pipeline A"
    }
  ],
  "components": {
    "schemas": {
...
      "AgreementEnvelope": {
        "type": "object",
        "required": [
          "schema",
          "key_id",
          "tuple",
          "signature"
        ],
        "description": "Ed25519 signature over the RFC 8785 canonical JSON of the tuple. The envelope fields around the tuple are not covered by the signature.",
        "properties": {
          "schema": {
            "type": "string",
            "enum": [
              "urn:tandem:agreement-envelope"
            ]
          },
          "key_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9._:-]{1,128}$"
          },
          "signature": {
            "type": "string",
            "pattern": "^[0-9a-f]{128}$"
          },
          "tuple": {
            "type": "object",
            "required": [
              "schema",
              "protocol_id",
              "height",
              "block_hash",
              "event_root",
              "object_state_root",
              "chained_root",
              "founding_created",
              "all_objects",
              "active_objects",
              "parser_commit",
              "indexer_commit",
              "parser_binary_sha256",
              "indexer_binary_sha256"
            ],
            "properties": {
              "schema": {
                "type": "string",
                "enum": [
                  "urn:tandem:agreement-tuple"
                ]
              },
              "protocol_id": {
                "type": "string",
                "pattern": "^tndm:(mainnet|signet|testnet4|regtest):[0-9a-f]{64}$"
              },
              "height": {
                "type": "string",
                "pattern": "^(0|[1-9][0-9]*)$"
              },
              "block_hash": {
                "type": "string",
                "pattern": "^[0-9a-f]{64}$"
              },
              "event_root": {
                "type": "string",
                "pattern": "^[0-9a-f]{64}$"
              },
              "object_state_root": {
                "type": "string",
                "pattern": "^[0-9a-f]{64}$"
              },
              "chained_root": {
                "type": "string",
                "pattern": "^[0-9a-f]{64}$"
              },
              "founding_created": {
                "type": "string",
                "pattern": "^(0|[1-9][0-9]*)$"
              },
              "all_objects": {
                "type": "string",
                "pattern": "^(0|[1-9][0-9]*)$"
              },
              "active_objects": {
                "type": "string",
                "pattern": "^(0|[1-9][0-9]*)$"
              },
              "parser_commit": {
                "type": "string",
                "pattern": "^[0-9a-f]{40}$"
              },
              "indexer_commit": {
                "type": "string",
                "pattern": "^[0-9a-f]{40}$"
              },
              "parser_binary_sha256": {
                "type": "string",
                "pattern": "^[0-9a-f]{64}$"
              },
              "indexer_binary_sha256": {
                "type": "string",
                "pattern": "^[0-9a-f]{64}$"
              }
            }
          }
        }
      }
    }
  },
...
}